pyproject.toml
[project]
name = "mushroom-man"
version = "0.1.0"
description = "A mushroom man discord bot."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "disckit>=1.1.4",
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.45",
]

[dependency-groups]
dev = [
    "basedpyright>=1.36.2",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 79
lint.extend-select = ["ANN", "I"]
lint.ignore = ["ANN401"]
lint.fixable = ["I", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 72

[tool.basedpyright]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    "venv/",
    "build/",
    "dist/",
    "local/",
    "examples/",
    "docs/",
    "lib_testing/",
]
reportImportCycles = false
reportUnnecessaryIsInstance = false
reportImplicitStringConcatenation = false
reportUnusedCallResult = false
reportMissingTypeStubs = false
reportAny = false
reportExplicitAny = false
reportImplicitOverride = false
